refactor(chess): change package structure
