+++ src/main/java/chess/domain/game/Started.java
@@ -3,6 +3,7 @@ package chess.domain.game;
 import static chess.domain.piece.Color.BLACK;
 import static chess.domain.piece.Color.WHITE;
 
+import chess.domain.piece.ActivePieces;
 import chess.dto.GameResultDto;
 
 abstract class Started implements Game {
