+++ src/main/java/chess/domain/piece/Pawn.java
@@ -3,7 +3,6 @@ package chess.domain.piece;
 import static chess.domain.piece.Color.BLACK;
 import static chess.domain.piece.Color.WHITE;
 
-import chess.domain.position.Position;
 import chess.strategy.OccupiedChecker;
 import java.util.Objects;
