+++ src/main/java/chess/domain/piece/strongmen/Bishop.java
@@ -1,36 +1,36 @@
-package chess.domain.piece;
+package chess.domain.piece.strongmen;
 
-import chess.domain.position.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Position;
 import java.util.Objects;
 
-public final class Rook extends Strongmen {
+public final class Bishop extends Strongmen {
 
-    private static final int LEFT_INIT_FILE = 0;
-    private static final int RIGHT_INIT_FILE = 7;
+    private static final int LEFT_INIT_FILE = 2;
+    private static final int RIGHT_INIT_FILE = 5;
 
-    private static final double SCORE = 5;
+    private static final double SCORE = 3;
 
-    private static final String BLACK_DISPLAY = "♖";
-    private static final String WHITE_DISPLAY = "♜";
+    private static final String BLACK_DISPLAY = "♙";
+    private static final String WHITE_DISPLAY = "♟";
 
-    Rook(Color color, Position position) {
+    Bishop(Color color, Position position) {
         super(color, position);
     }
 
-    public static Rook ofLeft(Color color) {
+    public static Bishop ofLeft(Color color) {
         Position position = Position.of(LEFT_INIT_FILE, firstRankOf(color));
-        return new Rook(color, position);
+        return new Bishop(color, position);
     }
 
-    public static Rook ofRight(Color color) {
+    public static Bishop ofRight(Color color) {
         Position position = Position.of(RIGHT_INIT_FILE, firstRankOf(color));
-        return new Rook(color, position);
+        return new Bishop(color, position);
     }
 
     @Override
     protected boolean canMoveTo(Position targetPosition) {
-        return position.isHorizontal(targetPosition)
-                || position.isVertical(targetPosition);
+        return position.isDiagonal(targetPosition);
     }
 
     @Override
@@ -59,9 +59,9 @@ public final class Rook extends Strongmen {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Rook rook = (Rook) o;
-        return color == rook.color
-                && position == rook.position;
+        Bishop bishop = (Bishop) o;
+        return color == bishop.color
+                && position == bishop.position;
     }
 
     @Override
@@ -71,6 +71,6 @@ public final class Rook extends Strongmen {
 
     @Override
     public String toString() {
-        return "Rook{color=" + color + ", position=" + position + '}';
+        return "Bishop{color=" + color + ", position=" + position + '}';
     }
 }
