+++ src/main/java/chess/domain/piece/strongmen/Queen.java
@@ -1,6 +1,7 @@
-package chess.domain.piece;
+package chess.domain.piece.strongmen;
 
-import chess.domain.position.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Position;
 import java.util.Objects;
 
 public final class Queen extends Strongmen {
