+++ src/main/java/chess/domain/piece/strongmen/Rook.java
@@ -1,35 +1,37 @@
-package chess.domain.piece;
+package chess.domain.piece.strongmen;
 
-import chess.domain.position.Position;
+import chess.domain.piece.Color;
+import chess.domain.piece.Position;
 import java.util.Objects;
 
-public final class Bishop extends Strongmen {
+public final class Rook extends Strongmen {
 
-    private static final int LEFT_INIT_FILE = 2;
-    private static final int RIGHT_INIT_FILE = 5;
+    private static final int LEFT_INIT_FILE = 0;
+    private static final int RIGHT_INIT_FILE = 7;
 
-    private static final double SCORE = 3;
+    private static final double SCORE = 5;
 
-    private static final String BLACK_DISPLAY = "♙";
-    private static final String WHITE_DISPLAY = "♟";
+    private static final String BLACK_DISPLAY = "♖";
+    private static final String WHITE_DISPLAY = "♜";
 
-    Bishop(Color color, Position position) {
+    Rook(Color color, Position position) {
         super(color, position);
     }
 
-    public static Bishop ofLeft(Color color) {
+    public static Rook ofLeft(Color color) {
         Position position = Position.of(LEFT_INIT_FILE, firstRankOf(color));
-        return new Bishop(color, position);
+        return new Rook(color, position);
     }
 
-    public static Bishop ofRight(Color color) {
+    public static Rook ofRight(Color color) {
         Position position = Position.of(RIGHT_INIT_FILE, firstRankOf(color));
-        return new Bishop(color, position);
+        return new Rook(color, position);
     }
 
     @Override
     protected boolean canMoveTo(Position targetPosition) {
-        return position.isDiagonal(targetPosition);
+        return position.isHorizontal(targetPosition)
+                || position.isVertical(targetPosition);
     }
 
     @Override
@@ -58,9 +60,9 @@ public final class Bishop extends Strongmen {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Bishop bishop = (Bishop) o;
-        return color == bishop.color
-                && position == bishop.position;
+        Rook rook = (Rook) o;
+        return color == rook.color
+                && position == rook.position;
     }
 
     @Override
@@ -70,6 +72,6 @@ public final class Bishop extends Strongmen {
 
     @Override
     public String toString() {
-        return "Bishop{color=" + color + ", position=" + position + '}';
+        return "Rook{color=" + color + ", position=" + position + '}';
     }
 }
