+++ src/test/java/chess/domain/game/BlackTurnTest.java
@@ -5,7 +5,8 @@ import static chess.domain.piece.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.piece.King;
+import chess.domain.piece.ActivePieces;
+import chess.domain.piece.strongmen.King;
 import chess.domain.piece.Piece;
 import chess.dto.MoveCommandDto;
 import java.util.ArrayList;
