+++ src/test/java/chess/util/PieceGeneratorUtilTest.java
@@ -4,13 +4,13 @@ import static chess.domain.piece.Color.BLACK;
 import static chess.domain.piece.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.piece.Bishop;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
+import chess.domain.piece.strongmen.Bishop;
+import chess.domain.piece.strongmen.King;
+import chess.domain.piece.strongmen.Knight;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.strongmen.Queen;
+import chess.domain.piece.strongmen.Rook;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.IntStream;
